[MOD] : SavedArticleViewHolder 분리 #130
