+++ app/src/main/java/com/indipage/presentation/savedarticle/viewholder/SaveArticleViewHolder.kt
@@ -1,0 +1,29 @@
+package com.indipage.presentation.savedarticle.viewholder
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+import com.indipage.databinding.ItemSavedArticleBinding
+import com.indipage.domain.entity.Article
+
+class SavedArticleViewHolder(
+    private val binding: ItemSavedArticleBinding,
+    private val onMoveToArticleDetailClick: (Article, Int) -> Unit = { _, _ -> }
+) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(data: Article) {
+        binding.savedArticle = data
+        if (data.ticketReceived) {
+            binding.clTicketTrue.visibility = View.VISIBLE
+            binding.clTicketFalse.visibility = View.GONE
+        } else {
+            binding.clTicketTrue.visibility = View.GONE
+            binding.clTicketFalse.visibility = View.VISIBLE
+        }
+        binding.clTicketTrue.setOnClickListener {
+            onMoveToArticleDetailClick(data, position)
+        }
+        binding.clTicketFalse.setOnClickListener {
+            onMoveToArticleDetailClick(data, position)
+        }
+        binding.executePendingBindings()
+    }
+}
